Stabilize network tests and share tc rate formatter
